Move `Metadata` to `Cardano.Read.Ledger.Tx.Metadata`
