+++ lib/read/cardano-wallet-read.cabal
@@ -57,6 +57,7 @@ library
     Cardano.Read.Ledger.Tx.Fee
     Cardano.Read.Ledger.Tx.Hash
     Cardano.Read.Ledger.Tx.Integrity
+    Cardano.Read.Ledger.Tx.Metadata
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
     Cardano.Wallet.Read.Block.BHeader
@@ -89,7 +90,6 @@ library
     Cardano.Wallet.Read.Tx.Gen.Shelley
     Cardano.Wallet.Read.Tx.Gen.TxParameters
     Cardano.Wallet.Read.Tx.Inputs
-    Cardano.Wallet.Read.Tx.Metadata
     Cardano.Wallet.Read.Tx.Mint
     Cardano.Wallet.Read.Tx.ReferenceInputs
     Cardano.Wallet.Read.Tx.ScriptValidity
